Rework script evaluation so auto-completion works in IntelliJ IDEA
