+++ polyglot-kotlin/README.md
@@ -74,7 +74,7 @@ project("Polyglot :: Kotlin") {
         }
 
         // External execute task script
-        execute(id = "hello-script", phase = "process-resources", script = "src/build/kotlin/hello.kts")
+        execute(id = "hello-script", phase = "process-resources", script = "hello.task.kts")
     }
 }
 ```
