--- polyglot-kotlin/src/main/kotlin/org/sonatype/maven/polyglot/kotlin/engine/PomKtsScriptDefinition.kt
@@ -1,5 +1,0 @@
-package org.sonatype.maven.polyglot.kotlin.engine
-
-import org.sonatype.maven.polyglot.kotlin.dsl.DSL
-
-object PomKtsScriptDefinition : ScriptDefinition(DSL::class)
