+++ polyglot-kotlin/src/main/kotlin/org/sonatype/maven/polyglot/kotlin/engine/TaskKtsScript.kt
@@ -1,10 +1,14 @@
 package org.sonatype.maven.polyglot.kotlin.engine
 
+import org.sonatype.maven.polyglot.execute.ExecuteContext
+import java.io.File
 import kotlin.script.experimental.annotations.KotlinScript
+import kotlin.script.templates.ScriptTemplateDefinition
 
 @KotlinScript(
     displayName = "External kotlin task script",
     fileExtension = "task.kts",
-    compilationConfiguration = TaskKtsScriptDefinition::class
+    compilationConfiguration = ScriptDefinition::class
 )
-abstract class TaskKtsScript
+@ScriptTemplateDefinition(scriptFilePattern = ".*\\.task\\.kts")
+abstract class TaskKtsScript(val script: File, private val delegate: ExecuteContext) : ExecuteContext by delegate
