--- polyglot-kotlin/src/main/kotlin/org/sonatype/maven/polyglot/kotlin/engine/TaskKtsScriptDefinition.kt
@@ -1,5 +1,0 @@
-package org.sonatype.maven.polyglot.kotlin.engine
-
-import org.sonatype.maven.polyglot.kotlin.execute.TaskContext
-
-object TaskKtsScriptDefinition : ScriptDefinition(TaskContext::class)
